--- pkg/storage/stores/shipper/bloomshipper/store_test.go
@@ -1,11 +1,0 @@
-package bloomshipper
-
-import (
-	"testing"
-)
-
-func TestBloomShipper(_ *testing.T) {
-}
-
-func TestBloomStore(_ *testing.T) {
-}
